s390/appldata_net_sum: use net_get_dev_array when calling dev_get_stats
In the effort of making .ndo_get_stats64 be able to sleep, we need to
ensure the callers of dev_get_stats do not use atomic context.

The ops->callback function is called from under appldata_ops_mutex
protection, so this is proof that the context is sleepable and we can
make the appldata driver use the new API for reference counting network
device arrays.

Cc: Heiko Carstens <hca@linux.ibm.com>
Cc: Vasily Gorbik <gor@linux.ibm.com>
Cc: linux-s390@vger.kernel.org
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>