parisc/led: use net_get_dev_array when calling dev_get_stats
In the effort of making .ndo_get_stats64 be able to sleep, we need to
ensure the callers of dev_get_stats do not use atomic context.

The LED driver for HP-PARISC workstations uses a workqueue to
periodically check for updates in network interface statistics (only
looks for interfaces with an IPv4 configuration and which are not
loopback), and flicker when those have changed (i.e. there has been
activity on the line).

Ignoring the fact that this driver is completely duplicating
drivers/leds/trigger/ledtrig-netdev.c, there is an even bigger problem.
Now, the dev_get_stats call can sleep, and iterating through the list of
network interfaces with for_each_netdev_rcu loses the sleepable context
necessary for the callee. So we need to use the newly introduced API for
taking temporary references on the network interfaces, and calling
dev_get_stats from outside RCU read-lock protection.

This patch also reindents the code that gathers device statistics, since
the standard in the Linux kernel is to use one tab character per
indentation level.

Cc: "James E.J. Bottomley" <James.Bottomley@HansenPartnership.com>
Cc: Helge Deller <deller@gmx.de>
Cc: linux-parisc@vger.kernel.org
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>