net: mscc: ocelot: implement port flushing on .adjust_link
Especially with flow control enabled on both the user port and the CPU
port, it may happen when a link goes down that Ethernet packets are in
flight. In flow control mode, frames are held back and not dropped. When
there is enough traffic in flight (example: iperf3 TCP), then the
ingress port might enter congestion and never exit that state. This is a
problem, because it is the egress port's link that went down, and that
has caused the inability of the ingress port to send packets to any
other port.

The solution is to follow the port flushing procedure from the reference
manual. This ensures that upon detection of link loss, the existing
packets are thrown away and congestion on the ingress port is therefore
avoided.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>