net: bridge: notify switchdev only of changes in br_vlan_add_existing
The br_vlan_add_existing function is called when br_vlan_add, the
function to add a VLAN to the bridge device's VLAN group, encounters an
already present VLAN entry.

This function is supposed to detect when it makes any changes to that
VLAN. There are two possible changes it can make:
- The VLAN was created by nbp_vlan_add -> __vlan_add -> br_vlan_get_master
  to ensure that the VLAN from the port's VLAN group has a corresponding
  master VLAN to be used with VLAN statistics per port, etc (but not for
  filtering). Then, when the user runs "bridge vlan add dev br0 vid 100 self",
  this causes a change in the master VLAN in that it becomes available
  for filtering.
- Any subsequent "bridge vlan add dev br0 vid 100 untagged self" command
  will change the flags of this master VLAN, and will qualify as a change
  in br_vlan_add_existing.

The point is that consecutive and identical "bridge vlan add ..." commands
should not notify switchdev if nothing changes. So do the notifier call
only after we made sure that switchdev has something to actually learn
about.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>