net: convert netdev_lower_get_next_private to netif_lists_lock
The bonding driver is the only in-kernel user of netdev_lower_get_next_private,
via:

bond_for_each_slave
-> netdev_for_each_lower_private
   -> netdev_lower_get_next_private

For this one, the new locking rules needed to be propagated to all callers of
bond_for_each_slave, because that is just a macro over an iterator, we can't
hide a locking wrapper inside it.

As a somewhat nice side effect, it looks like some entry points for the
bonding driver, namely the sysfs interface, did only need the RTNL mutex
for iterating through the list of slaves. It doesn't need that any
longer.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>