net: convert netdev_walk_all_lower_dev to netif_lists_lock
netdev_walk_all_lower_dev has no direct users in net/core/dev.c, and the
switchdev drivers that call this can all be oblivious of the fact that
protection is necessary when traversing the netif adjacency list. So
take the necessary protection privately in netdev_walk_all_lower_dev.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>