net: convert netdev_{,master_}upper_dev_link to netif_lists_lock
The callers of netdev_master_upper_dev_link and netdev_upper_dev_link
still need to hold the RTNL mutex because of the netdevice notifiers
emitted by this function: NETDEV_PRECHANGEUPPER and NETDEV_CHANGEUPPER.

Because of the strict ordering requirement between the rtnl_mutex and
the netif_lists_lock (the rtnl_mutex is required to be top-level), plus
the fact that both locks are required in this function, the effect is
that the callers of netdev_upper_dev_link cannot be aware of
netif_lists_lock at all. This function needs to take it privately, and
this is what this patch does. The locking protects the network interface
adjacency lists.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>