net: add ASSERT_RTNL to netdev_adjacent_change_{prepare,commit,abort}
The transactional implementation of rtnl_link_ops::changelink that the
VXLAN driver has comes from commit 0ce1822c2a08 ("vxlan: add adjacent
link to limit depth level"). The goal there was to add a limit to the
nesting level of VXLAN interfaces.

Under the hood, the "prepare" phase of a netif adjacency change actually
does the real thing, except that it sets ignore=true. Therefore the same
__netdev_upper_dev_link is called, and the netdev notifiers are called
too. So the rtnl_mutex is still required. Make that more obvious by
slapping an ASSERT_RTNL in there.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>